Separate CreateEnvironmentDropdown
