+++ packages/webapp/src/components-v2/AppSidebar/EnvironmentDropdown.tsx
@@ -1,22 +1,16 @@
-import { zodResolver } from '@hookform/resolvers/zod';
-import { Check, ChevronsUpDown, Loader, Lock } from 'lucide-react';
+import { Check, ChevronsUpDown, Lock } from 'lucide-react';
 import React, { useState } from 'react';
-import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
-import { z } from 'zod';
 
 import { StyledLink } from '../StyledLink.js';
+import { CreateEnvironmentDialog } from './CreateEnvironmentDialog.js';
 import { Button } from '../ui/button.js';
-import { Dialog, DialogClose, DialogContent, DialogFooter, DialogHeader, DialogTitle } from '../ui/dialog.js';
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '../ui/dropdown-menu.js';
-import { Form, FormControl, FormDescription, FormField, FormItem, FormMessage } from '../ui/form.js';
-import { Input } from '../ui/input.js';
 import { SidebarMenu, SidebarMenuItem } from '../ui/sidebar.js';
 import { Tooltip, TooltipContent, TooltipTrigger } from '../ui/tooltip.js';
 import { LogoInverted } from '@/assets/LogoInverted';
-import { apiPostEnvironment, useEnvironment } from '@/hooks/useEnvironment';
+import { useEnvironment } from '@/hooks/useEnvironment';
 import { useMeta } from '@/hooks/useMeta';
-import { useToast } from '@/hooks/useToast';
 import { useStore } from '@/store';
 
 export const EnvironmentDropdown: React.FC = () => {
@@ -25,7 +19,6 @@ export const EnvironmentDropdown: React.FC = () => {
     const envs = useStore((state) => state.envs);
     const environment = useEnvironment(env);
     const { meta } = useMeta();
-    const [dropdownMenuOpen, setDropdownMenuOpen] = useState(false);
     const [environmentDialogOpen, setEnvironmentDialogOpen] = useState(false);
 
     const navigate = useNavigate();
@@ -62,7 +55,7 @@ export const EnvironmentDropdown: React.FC = () => {
     return (
         <SidebarMenu>
             <SidebarMenuItem>
-                <DropdownMenu open={dropdownMenuOpen} onOpenChange={setDropdownMenuOpen} modal={false}>
+                <DropdownMenu modal={false}>
                     <DropdownMenuTrigger className="h-fit w-full rounded p-2.5 flex flex-row items-center justify-between cursor-pointer bg-dropdown-bg-default border-[0.5px] border-border-muted hover:bg-dropdown-bg-press hover:border-0 hover:border-l-[0.5px] hover:border-r-[0.5px] hover:border-r-transparent data-[state=closed]:hover:my-[0.5px] data-[state=open]:bg-dropdown-bg-press data-[state=open]:border-[0.5px] data-[state=open]:border-border-muted">
                         <div className="flex gap-2 items-center">
                             <LogoInverted className="h-6 w-6 text-text-primary" />
@@ -79,7 +72,7 @@ export const EnvironmentDropdown: React.FC = () => {
                                 key={environment.name}
                                 onSelect={() => onSelect(environment.name)}
                                 data-active={env === environment.name}
-                                className="flex flex-row items-center gap-2 data-[active=true]:text-text-primary"
+                                className="flex flex-row items-center gap-2 cursor-pointer data-[active=true]:text-text-primary"
                             >
                                 <Check
                                     className="w-5 h-5 opacity-0 data-[active=true]:opacity-100 data-[active=true]:text-text-primary"
@@ -94,8 +87,6 @@ export const EnvironmentDropdown: React.FC = () => {
                                     <Button
                                         variant="primary"
                                         onClick={() => {
-                                            // We have to close the dropdown because it traps focus
-                                            setDropdownMenuOpen(false);
                                             // Managed control because Dialogs within DropdownMenus behave weirdly
                                             setEnvironmentDialogOpen(true);
                                         }}
@@ -128,93 +119,5 @@ export const EnvironmentDropdown: React.FC = () => {
                 <CreateEnvironmentDialog open={environmentDialogOpen} onOpenChange={setEnvironmentDialogOpen} />
             </SidebarMenuItem>
         </SidebarMenu>
-    );
-};
-
-interface CreateEnvironmentDialogProps {
-    open: boolean;
-    onOpenChange: (open: boolean) => void;
-}
-
-const environmentFormSchema = z.object({
-    name: z
-        .string()
-        .regex(/^[a-z0-9_-]+$/, 'Invalid environment name')
-        .max(255)
-});
-
-type EnvironmentForm = z.infer<typeof environmentFormSchema>;
-
-export const CreateEnvironmentDialog: React.FC<CreateEnvironmentDialogProps> = ({ open, onOpenChange }) => {
-    const form = useForm<EnvironmentForm>({
-        resolver: zodResolver(environmentFormSchema),
-        defaultValues: {
-            name: ''
-        }
-    });
-
-    const { toast } = useToast();
-    const navigate = useNavigate();
-    const { mutate: mutateMeta } = useMeta();
-
-    const [loading, setLoading] = useState(false);
-
-    async function onSubmit(data: EnvironmentForm) {
-        setLoading(true);
-
-        const res = await apiPostEnvironment({ name: data.name });
-        if ('error' in res.json) {
-            const err = res.json.error;
-            if (err.code === 'invalid_body') {
-                form.setError('name', { message: 'Invalid environment name' });
-            } else if (['conflict', 'feature_disabled', 'resource_capped'].includes(err.code)) {
-                toast({ title: err.message, variant: 'error' });
-            } else {
-                toast({ title: 'Failed to create environment', variant: 'error' });
-            }
-        } else {
-            navigate(`/${res.json.data.name}`);
-            onOpenChange(false);
-            form.reset();
-            void mutateMeta();
-        }
-
-        setLoading(false);
-    }
-
-    return (
-        <Dialog open={open} onOpenChange={onOpenChange} modal={true}>
-            <DialogContent>
-                <Form {...form}>
-                    <form onSubmit={form.handleSubmit(onSubmit)} className="flex flex-col gap-10">
-                        <DialogHeader>
-                            <DialogTitle>Environment Name</DialogTitle>
-                        </DialogHeader>
-                        <FormField
-                            control={form.control}
-                            name="name"
-                            render={({ field }) => (
-                                <FormItem>
-                                    <FormControl>
-                                        <Input type="text" placeholder="my-environment-name" {...field} />
-                                    </FormControl>
-                                    <FormDescription>*Must be lowercase letters, numbers, underscores and dashes.</FormDescription>
-                                    <FormMessage />
-                                </FormItem>
-                            )}
-                        />
-                        <DialogFooter>
-                            <DialogClose asChild>
-                                <Button variant="secondary">Cancel</Button>
-                            </DialogClose>
-                            <Button variant="primary" type="submit" disabled={loading}>
-                                {loading && <Loader className="animate-spin h-full w-full" />}
-                                Create Environment
-                            </Button>
-                        </DialogFooter>
-                    </form>
-                </Form>
-            </DialogContent>
-        </Dialog>
     );
 };
